Create a text file with all NOTAM

+++ install/class.update_schema.php
@@ -946,11 +946,9 @@ class update_schema {
 				return "error (add stats_airline & filter_name column in stats) : ".$e->getMessage()."\n";
 			}
 		}
-		if ($Connection->indexExists('stats','type')) {
+		if ($globalDBdriver == 'mysql' && $Connection->indexExists('stats','type')) {
 			// Add unique key
-			if ($globalDBdriver == 'mysql') {
-				$query = "drop index type on stats;ALTER TABLE stats ADD UNIQUE stats_type (stats_type,stats_date,stats_airline,filter_name);";
-			}
+			$query = "drop index type on stats;ALTER TABLE stats ADD UNIQUE stats_type (stats_type,stats_date,stats_airline,filter_name);";
 			try {
 				$sth = $Connection->db->prepare($query);
 				$sth->execute();
